Quit the game automatically when a non existing world is added.

+++ Assets/Scripts/WorldManager.cs
@@ -154,6 +154,9 @@ public class WorldManager : MonoBehaviour
 
     public void deleteWorld ()
     {
-        world.delete();
+        if (world != null)
+        {
+            world.delete();
+        }
     }
 }
